+++ tile-rendering/src/main/java/com/oculusinfo/tile/rendering/impl/NumberListHeatMapImageRenderer.java
@@ -82,8 +82,8 @@ public class NumberListHeatMapImageRenderer implements TileDataImageRenderer<Lis
 		try {
 			int outputWidth = config.getPropertyValue( LayerConfiguration.OUTPUT_WIDTH );
 			int outputHeight = config.getPropertyValue( LayerConfiguration.OUTPUT_HEIGHT );
-			double rangeMax = config.getPropertyValue( LayerConfiguration.RANGE_MAX );
-			double rangeMin = config.getPropertyValue( LayerConfiguration.RANGE_MIN );
+			int rangeMax = config.getPropertyValue( LayerConfiguration.RANGE_MAX );
+			int rangeMin = config.getPropertyValue( LayerConfiguration.RANGE_MIN );
 			String rangeMode = config.getPropertyValue( LayerConfiguration.RANGE_MODE );
 			String pixelShape = config.getPropertyValue( LayerConfiguration.PIXEL_SHAPE );
 
@@ -93,7 +93,6 @@ public class NumberListHeatMapImageRenderer implements TileDataImageRenderer<Lis
 			ValueTransformer<Number> t = config.produce( "valueTransformer", ValueTransformer.class );
 			@SuppressWarnings( "unchecked" )
 			ValueTransformer<Number> alphaTransformer = config.produce( "alphaValueTransformer", ValueTransformer.class );
-
 			double scaledMax = ( double ) rangeMax / 100;
 			double scaledMin = ( double ) rangeMin / 100;
